+++ Braintree/src/main/java/com/braintreepayments/api/ThreeDSecure.java
@@ -49,7 +49,7 @@ public class ThreeDSecure {
     /**
      * The versioned path of the 3D Secure assets to use. Hosted by Braintree.
      */
-    private static final String THREE_D_SECURE_ASSETS_PATH = "/mobile/three-d-secure-redirect/0.1.5";
+    private static final String THREE_D_SECURE_ASSETS_PATH = "/mobile/three-d-secure-redirect/0.1.6";
 
     /**
      * @deprecated Use {{@link #performVerification(BraintreeFragment, CardBuilder, ThreeDSecureRequest)}} for 3DS 2.0.